refactor: rename the provider name to a proper name
