+++ src/index.js
@@ -3,13 +3,13 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import NewProvider from './context/NewProvider';
+import NewsProvider from './context/NewsProvider';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <NewProvider>
+  <NewsProvider>
     <App />
-  </NewProvider>
+  </NewsProvider>
 );
 
 // If you want to start measuring performance in your app, pass a function
